Move new configs outside of 'custom' package - Create properties package for storing the config properties - Move NewSetting class to main settings pkg
