+++ src/test/java/fr/xephi/authme/settings/NewSettingIntegrationTest.java
@@ -1,8 +1,10 @@
-package fr.xephi.authme.settings.custom;
+package fr.xephi.authme.settings;
 
 import com.google.common.collect.ImmutableMap;
 import fr.xephi.authme.ReflectionTestUtils;
+import fr.xephi.authme.settings.NewSetting;
 import fr.xephi.authme.settings.domain.Property;
+import fr.xephi.authme.settings.properties.TestConfiguration;
 import fr.xephi.authme.settings.propertymap.PropertyMap;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.junit.BeforeClass;
