+++ src/test/java/fr/xephi/authme/settings/NewSettingTest.java
@@ -1,6 +1,8 @@
-package fr.xephi.authme.settings.custom;
+package fr.xephi.authme.settings;
 
+import fr.xephi.authme.settings.NewSetting;
 import fr.xephi.authme.settings.domain.Property;
+import fr.xephi.authme.settings.properties.TestConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.junit.Test;
 import org.mockito.invocation.InvocationOnMock;
